third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger/script-failed-to-parse-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that ParsedScriptSource event is raised after compile script with syntax error.


Running: testScriptParsedEvent
Event with script-failed-to-parse.js received
